requirements.txt
fastapi
uvicorn[standard]
firebase-admin
requests
python-multipart
pillow
cloudinary
openai
gTTS
vosk
pydantic
tensorflow
pandas
torch
torchvision